+++ app/views/frontend/modules/product/new_product.blade.php
@@ -47,7 +47,7 @@
 											<select required="required" class="form-control" name="s_category">
 												<?php foreach($categoryTree as $cl) : ?>
 													<option value="{{$cl['id']}}">
-														<?php if (Session::get('currentUserAccountType') === 1): ?>
+														<?php if ((int)Session::get('currentUserAccountType') === 1): ?>
 															{{$cl['name_en']}}
 														<?php else: ?>
 															{{$cl['m_title']}}
